% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_moments.R
\name{check_input}
\alias{check_input}
\alias{check_input.default}
\alias{check_input.numeric}
\alias{check_input.matrix}
\alias{check_input.xts}
\alias{check_input.tbl_df}
\title{Internal function used to check the validity of inputs.}
\usage{
check_input(x)

\method{check_input}{default}(x)

\method{check_input}{numeric}(x)

\method{check_input}{matrix}(x)

\method{check_input}{xts}(x)

\method{check_input}{tbl_df}(x)
}
\arguments{
\item{x}{Any object passed to other functions in the package.}
}
\value{
A matrix
}
\description{
Internal function used to check the validity of inputs.
}
\keyword{internal}
