% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_moments.R
\name{check_p}
\alias{check_p}
\alias{check_p.default}
\alias{check_p.NULL}
\alias{check_p.numeric}
\alias{check_p.ffp}
\alias{check_p.matrix}
\alias{check_p.xts}
\alias{check_p.data.frame}
\alias{check_p.tbl}
\title{Internal functions to check the consistency of probabilities.}
\usage{
check_p(p)

\method{check_p}{default}(p)

\method{check_p}{`NULL`}(p)

\method{check_p}{numeric}(p)

\method{check_p}{ffp}(p)

\method{check_p}{matrix}(p)

\method{check_p}{xts}(p)

\method{check_p}{data.frame}(p)

\method{check_p}{tbl}(p)
}
\arguments{
\item{p}{A vector, a matrix, a xts or a tibble object.}
}
\value{
A matrix object with 1 column.
}
\description{
Internal functions to check the consistency of probabilities.
}
\examples{
#
}
