% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{data}
\name{db}
\alias{db}
\title{Dataset used in Historical Scenarios with Fully Flexible Probabilities
(\code{matrix} format).}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1083 rows and 9 columns.
}
\usage{
data(db)
}
\description{
Dataset used in Historical Scenarios with Fully Flexible Probabilities
(\code{matrix} format).
}
\seealso{
\code{\link{db_tbl}}
}
\keyword{datasets}
