% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\docType{data}
\name{db_tbl}
\alias{db_tbl}
\title{Dataset used in Historical Scenarios with Fully Flexible Probabilities
(\code{tibble} format).}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1083 rows and 9 columns.
}
\usage{
data(db_tbl)
}
\description{
Dataset used in Historical Scenarios with Fully Flexible Probabilities
(\code{tibble} format).
}
\seealso{
\code{\link{db}}
}
\keyword{datasets}
