% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Double_Decay.R
\name{DoubleDecay}
\alias{DoubleDecay}
\title{Double-Decay Covariance Matrix}
\usage{
DoubleDecay(x, decay_low, decay_high)
}
\arguments{
\item{x}{A numeric matrix containing the relevant risk drivers.}

\item{decay_low}{A numeric decay rate used to estimate correlations.}

\item{decay_high}{A numeric decay rate used to estimate volatilities.}
}
\value{
A list containing the target posterior mean and covariance matrix.
}
\description{
This function computes the covariance matrix using two different decay
factors.
}
\details{
A common practice is to estimate covariance using a slower decay for
correlations and a faster decay for volatilities.
}
\keyword{internal}
