% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens.R
\name{ens}
\alias{ens}
\title{Effective Number of Scenarios}
\usage{
ens(p)
}
\arguments{
\item{p}{An object of class \code{ffp}.}
}
\value{
A single \code{double}.
}
\description{
Computes the effective number of scenarios implied by a flexible probability
distribution.
}
\details{
The effective number of scenarios is defined as the exponential of the
Shannon entropy:

\deqn{
ENS = \exp\left(-\sum_t p_t \log p_t\right).
}

Probabilities equal to zero make no contribution to the entropy, following
the convention \eqn{0 \log(0) = 0}.
}
\examples{
p <- as_ffp(rep(0.01, 100))
ens(p)

p <- as_ffp(c(0.5, 0.5, 0))
ens(p)
}
