% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_information.R
\name{exp_smoothing}
\alias{exp_smoothing}
\alias{exp_smoothing.default}
\alias{exp_smoothing.numeric}
\alias{exp_smoothing.matrix}
\alias{exp_smoothing.ts}
\alias{exp_smoothing.xts}
\alias{exp_smoothing.data.frame}
\alias{exp_smoothing.tbl}
\title{Full Information by Exponential Smoothing}
\usage{
exp_smoothing(x, lambda)

\method{exp_smoothing}{default}(x, lambda)

\method{exp_smoothing}{numeric}(x, lambda)

\method{exp_smoothing}{matrix}(x, lambda)

\method{exp_smoothing}{ts}(x, lambda)

\method{exp_smoothing}{xts}(x, lambda)

\method{exp_smoothing}{data.frame}(x, lambda)

\method{exp_smoothing}{tbl}(x, lambda)
}
\arguments{
\item{x}{An univariate or a multivariate distribution.}

\item{lambda}{A number for the decay parameter.}
}
\value{
A numerical vector of class \code{ffp} with the new
probabilities distribution.
}
\description{
Exponential smoothing twists probabilities by giving relatively more weight
to recent observations at an exponential rate.
}
\details{
The half-life is linked with the lambda parameter as follows:
\itemize{
\item \code{HL = log(2) / lambda}.
}

For example: log(2) / 0.0166 is approximately 42. So, a parameter \code{lambda} of 0.0166 can be associated with a half-life of two-months.
}
\examples{
library(ggplot2)

# long half_life
long_hl <- exp_smoothing(EuStockMarkets, 0.001)
long_hl
autoplot(long_hl) +
  scale_color_viridis_c()

# short half_life
short_hl <- exp_smoothing(EuStockMarkets, 0.015)
short_hl
autoplot(short_hl) +
  scale_color_viridis_c()
}
\seealso{
\code{\link{crisp}} \code{\link{kernel_normal}} \code{\link{half_life}}
}
