% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-imports.R
\docType{package}
\name{ffp-package}
\alias{ffp-package}
\title{Import vctrs generics for S3 methods}
\description{
Implements numerical entropy-pooling for portfolio construction and scenario analysis as described in Meucci, Attilio (2008) and Meucci, Attilio (2010) \doi{10.2139/ssrn.1696802}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Reckziegel/FFP}
  \item Report bugs at \url{https://github.com/Reckziegel/FFP/issues}
}

}
\author{
\strong{Maintainer}: Bernardo Reckziegel \email{bernardo_cse@hotmail.com}

Authors:
\itemize{
  \item Bernardo Reckziegel \email{bernardo_cse@hotmail.com}
}

}
\keyword{internal}
