% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp-vctrs.R
\name{ffp-vctrs}
\alias{ffp-vctrs}
\alias{new_ffp}
\alias{vec_ptype_abbr.ffp}
\alias{vec_ptype2.ffp.ffp}
\alias{vec_ptype2.ffp.double}
\alias{vec_ptype2.double.ffp}
\alias{vec_cast.ffp.ffp}
\alias{vec_cast.ffp.double}
\alias{vec_cast.double.ffp}
\alias{obj_print_data.ffp}
\alias{vec_math.ffp}
\alias{vec_arith.ffp}
\alias{vec_arith.numeric.ffp}
\title{Internal vctrs methods}
\usage{
new_ffp(x = double(), ...)

\method{vec_ptype_abbr}{ffp}(x, ...)

\method{vec_ptype2}{ffp.ffp}(x, y, ...)

\method{vec_ptype2}{ffp.double}(x, y, ...)

\method{vec_ptype2}{double.ffp}(x, y, ...)

\method{vec_cast}{ffp.ffp}(x, to, ...)

\method{vec_cast}{ffp.double}(x, to, ...)

\method{vec_cast}{double.ffp}(x, to, ...)

\method{obj_print_data}{ffp}(x, ...)

\method{vec_math}{ffp}(.fn, .x, ...)

\method{vec_arith}{ffp}(op, x, y, ...)

\method{vec_arith.numeric}{ffp}(op, x, y, ...)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
No return value, called for side effects.
}
\description{
Internal vctrs methods
}
\keyword{internal}
