% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_ffp_tbl}
\alias{new_ffp_tbl}
\alias{validate_ffp_tbl}
\alias{ffp_tbl}
\alias{is_ffp_tbl}
\alias{as_ffp_tbl}
\title{Internal ffp_tbl class}
\usage{
new_ffp_tbl(x, ...)

validate_ffp_tbl(x)

ffp_tbl(x)

is_ffp_tbl(x)

as_ffp_tbl(x)
}
\description{
Internal ffp_tbl class
}
\keyword{internal}
