% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit2Moms.R
\name{Fit2Moms}
\alias{Fit2Moms}
\title{Double-Decay Covariance Matrix by Entropy-Pooling}
\usage{
Fit2Moms(X, m, S)
}
\arguments{
\item{X}{A \code{matrix} of size \code{TxN} with relevant the risk drivers.}

\item{m}{A \code{matrix} of size \code{Nx1} given by \code{\link{DoubleDecay}}
with risk drivers location.}

\item{S}{A \code{matrix} of size \code{NxN} given by \code{\link{DoubleDecay}}
with the dispersion matrix of the risk drivers.}
}
\value{
A \code{matrix} vector with the posterior probabilities.
}
\description{
Internal function that computes the flexible probabilities given by
double-decay covariance matrix.
}
\keyword{internal}
