% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_moments.R
\name{fit_to_moments}
\alias{fit_to_moments}
\title{Double-Decay Covariance Matrix by Entropy-Pooling}
\usage{
fit_to_moments(X, m, S)
}
\arguments{
\item{X}{A numeric matrix of size \verb{T x N} containing the scenario features.}

\item{m}{A numeric vector or matrix of length \code{N} containing the target
posterior means.}

\item{S}{A numeric matrix of size \verb{N x N} containing the target posterior
covariance matrix.}
}
\value{
A numeric vector containing the posterior probabilities.
}
\description{
Computes flexible probabilities that match target first and second moments
while minimizing relative entropy from a uniform prior.
}
\keyword{internal}
