% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{half_life}
\alias{half_life}
\title{Half-Life Calculation}
\usage{
half_life(lambda)
}
\arguments{
\item{lambda}{A number.}
}
\value{
A single number with the half-life in days.
}
\description{
Compute the implied half-life of a decay parameter.
}
\examples{
half_life(0.0166)
half_life(0.01)
}
\seealso{
\code{\link{exp_decay}} \code{\link{double_decay}}
}
