% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_information.R
\name{kernel_normal}
\alias{kernel_normal}
\alias{kernel_normal.default}
\alias{kernel_normal.numeric}
\alias{kernel_normal.matrix}
\alias{kernel_normal.ts}
\alias{kernel_normal.xts}
\alias{kernel_normal.tbl_df}
\alias{kernel_normal.data.frame}
\title{Full Information by Kernel-Damping}
\usage{
kernel_normal(x, mean, sigma)

\method{kernel_normal}{default}(x, mean, sigma)

\method{kernel_normal}{numeric}(x, mean, sigma)

\method{kernel_normal}{matrix}(x, mean, sigma)

\method{kernel_normal}{ts}(x, mean, sigma)

\method{kernel_normal}{xts}(x, mean, sigma)

\method{kernel_normal}{tbl_df}(x, mean, sigma)

\method{kernel_normal}{data.frame}(x, mean, sigma)
}
\arguments{
\item{x}{The risk-drivers.}

\item{mean}{The point in which the kernel should be centered.}

\item{sigma}{A number with the uncertainty (volatility) around the mean.}
}
\value{
A \code{tibble} with the new probabilities distribution.
}
\description{
In this framework, historical realizations receive a weight proportional to
it's distance from the target mean, which is enveloped by
normal kernel with a bandwidth equal sigma.
}
\examples{
library(ggplot2)

ret <- diff(log(EuStockMarkets[ , 1]))
mean <- -0.01 # scenarios around -1\%
sigma <- var(diff(ret))

kn <- kernel_normal(ret, mean, sigma)
kn

ggplot(kn, aes(x = .rowid, y = .p, color = .p)) +
  geom_line(show.legend = FALSE) +
  scale_color_viridis_c()

# Spread the distribution with a larger sigma
sigma <- var(diff(ret)) / 0.05
kn <- kernel_normal(ret, mean, sigma)

ggplot(kn, aes(x = .rowid, y = .p, color = .p)) +
  geom_line(show.legend = FALSE) +
  scale_color_viridis_c()
}
\seealso{
\code{\link{crisp}} \code{\link{smoothing}}
}
