% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_info_kernel.R
\name{least_info_kernel}
\alias{least_info_kernel}
\title{Least Information Kernel-Smoothing}
\usage{
least_info_kernel(Y, y, h2)
}
\arguments{
\item{Y}{A numeric matrix containing the scenario features.}

\item{y}{A numeric vector containing the target posterior means.}

\item{h2}{A numeric covariance matrix. If \code{NULL}, only the posterior means
are constrained.}
}
\value{
A numeric vector containing the posterior probabilities.
}
\description{
This is an internal function that uses the kernel-smoothing approach to
compute posterior probabilities satisfying target first and second moments.
}
\keyword{internal}
