% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastInfoKernel.R
\name{LeastInfoKernel}
\alias{LeastInfoKernel}
\title{Least Information Kernel-Smoothing}
\usage{
LeastInfoKernel(Y, y, h2)
}
\arguments{
\item{Y}{A \code{matrix} with the macroeconomic indicators.}

\item{y}{A \code{numeric} scalar with the target to which \code{Y} is expected
to be near by.}

\item{h2}{A \code{matrix} with the covariance of the \code{Y} factor.}
}
\value{
A \code{matrix} of size \code{TX1} with flexible probabilities.
}
\description{
This is an internal function that uses the kernel-smoothing approach to compute
the posterior probabilities that satisfies some macroeconomic statement.
}
\keyword{internal}
