% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_density.R
\name{scenario_density}
\alias{scenario_density}
\title{Plot Scenario Densities}
\usage{
scenario_density(x, p)
}
\arguments{
\item{x}{An univariate marginal distribution}

\item{p}{A vector of fully flexible probabilities.}
}
\value{
A \code{ggplot2} object.
}
\description{
SSSS.
}
\examples{
ret <- diff(log(EuStockMarkets))[, 1]
p <- smoothing(ret, 0.005)
scenario_density(ret, p)
}
