% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_information.R
\name{smoothing}
\alias{smoothing}
\alias{smoothing.default}
\alias{smoothing.numeric}
\alias{smoothing.matrix}
\alias{smoothing.ts}
\alias{smoothing.xts}
\alias{smoothing.data.frame}
\alias{smoothing.tbl}
\title{Full Information by Exponential Smoothing}
\usage{
smoothing(x, lambda)

\method{smoothing}{default}(x, lambda)

\method{smoothing}{numeric}(x, lambda)

\method{smoothing}{matrix}(x, lambda)

\method{smoothing}{ts}(x, lambda)

\method{smoothing}{xts}(x, lambda)

\method{smoothing}{data.frame}(x, lambda)

\method{smoothing}{tbl}(x, lambda)
}
\arguments{
\item{x}{A vector of risk-drivers.}

\item{lambda}{A number with a half-life parameter.}
}
\value{
A vector with the new probabilities distribution.
}
\description{
Exponential smoothing twists probabilities by giving relatively more weight
to recent observations at an exponential rate.
}
\examples{
ffp1 <- smoothing(EuStockMarkets, 0.015)
ffp2 <- smoothing(EuStockMarkets, 0.01)
ffp3 <- smoothing(EuStockMarkets, 0.0075)
ffp4 <- smoothing(EuStockMarkets, 0.005)

plot(ffp1, type = 'l')
lines(ffp2, type = 'l', col = 'red')
lines(ffp3, type = 'l', col = 'blue')
lines(ffp4, type = 'l', col = 'green')
}
