% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_by_rank.R
\name{view_by_rank}
\alias{view_by_rank}
\title{Input Relative Views}
\usage{
view_by_rank(.invariant, .p, .rank)
}
\arguments{
\item{.invariant}{A time series defining the scenario-probability distribution.}

\item{.p}{An object of the \code{ffp} class.}

\item{.rank}{A \code{double} with the asset indexes.}
}
\value{
A matrix.
}
\description{
Compute the relative over/under performance of assets.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)
colnames(x) <- colnames(EuStockMarkets)
prior <- as.matrix(rep(1/nrow(x), nrow(x)))
# asset in the first col will outperform the asset in the second col.
view_by_rank(.invariant = x, .p = prior, .rank = c(2, 1))
}
