% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_tail_dependence}
\alias{view_on_tail_dependence}
\alias{view_on_tail_dependence.default}
\alias{view_on_tail_dependence.matrix}
\alias{view_on_tail_dependence.xts}
\alias{view_on_tail_dependence.tbl_df}
\title{Views on Tail Dependence}
\usage{
view_on_tail_dependence(x, tail)

\method{view_on_tail_dependence}{default}(x, tail)

\method{view_on_tail_dependence}{matrix}(x, tail)

\method{view_on_tail_dependence}{xts}(x, tail)

\method{view_on_tail_dependence}{tbl_df}(x, tail)
}
\arguments{
\item{x}{A multivariate copula.}

\item{tail}{A \code{double} with tail index of each asset.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct views on tail dependence.
}
\examples{
library(ggplot2)
set.seed(1)

# Invariants
x <- diff(log(EuStockMarkets))
u <- apply(x, 2, stats::pnorm) # normal copula (assumption, it could be anything)
n <- nrow(x)
tail <- c(0.49, 0.5, 0.5002, 0.5) # tail index

# Prior Probabilities
prior <- rep(1 / n, n)

# Views
views <- view_on_tail_dependence(x = u, tail = tail)
views

# Optimization
ep <- entropy_pooling(p = prior, Aeq = views$Aeq, beq = views$beq, solver = "nlminb")
autoplot(ep)
}
