% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_probs.R
\name{bind_probs}
\alias{bind_probs}
\title{Stack Flexible-Probabilities}
\usage{
bind_probs(...)
}
\arguments{
\item{...}{Different \code{tibble} objects with the \code{ffp} signature to combine.}
}
\value{
A tidy \code{tibble}.
}
\description{
This function mimics \code{dplyr} \code{\link[dplyr]{bind}}. It's useful if you
have different \code{ffp} objects and want to stack them in the \code{tidy} (long) format.
}
\examples{
x <- smoothing(EuStockMarkets, 0.001)
y <- smoothing(EuStockMarkets, 0.002)
bind_probs(x, y)
}
\seealso{
\code{\link{crisp}} \code{\link{smoothing}} \code{\link{kernel_normal}}
\code{\link{kernel_entropy}} \code{\link{double_decay}}
}
