% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{bind_views}
\alias{bind_views}
\title{Stack Different Views}
\usage{
bind_views(...)
}
\arguments{
\item{...}{Objects of the class \code{ffp_views} to combine.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Bind views for entropy programming.
}
\examples{
library(ggplot2)

# Invariant
ret <- diff(log(EuStockMarkets))
n   <- nrow(ret)

# Prior belief for expected returns (here is 1\% for each asset)
mean <- rep(0, 4)

# Prior probabilities (usually equal weight scheme)
prior <- rep(1 / n, n)

# View on returns
view_mean <- view_on_mean(x = ret, mean = mean)

#' view on volatility
vol <- apply(ret, 2, stats::sd) * 1.1
view_volatility <- view_on_volatility(x = ret, vol = vol)

views_comb <- bind_views(view_mean, view_volatility)
views_comb

ep <- entropy_pooling(p      = prior,
                      Aeq    = views_comb$Aeq,
                      beq    = views_comb$beq,
                      A      = views_comb$A,
                      b      = views_comb$b,
                      solver = "nlminb")
autoplot(ep)
}
