% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_information.R
\name{crisp}
\alias{crisp}
\alias{crisp.default}
\alias{crisp.numeric}
\alias{crisp.matrix}
\alias{crisp.ts}
\alias{crisp.xts}
\alias{crisp.data.frame}
\alias{crisp.tbl_df}
\title{Full Information by Market Conditioning}
\usage{
crisp(x, lgl)

\method{crisp}{default}(x, lgl)

\method{crisp}{numeric}(x, lgl)

\method{crisp}{matrix}(x, lgl)

\method{crisp}{ts}(x, lgl)

\method{crisp}{xts}(x, lgl)

\method{crisp}{data.frame}(x, lgl)

\method{crisp}{tbl_df}(x, lgl)
}
\arguments{
\item{x}{A vector of risk-drivers.}

\item{lgl}{A \code{logical} vector with TRUE's and FALSE's indicating
which scenarios should considered.}
}
\value{
A \code{tibble} with the new probabilities distribution.
}
\description{
In this function probabilities are given full weight when a certain macroeconomic
condition satisfy a logical statement.
}
\examples{
# invariance (stationarity)
ret <- diff(log(EuStockMarkets))[ , 1]

# full weight on scenarios above 2\%
crisp(x = ret, ret > 0.02)
}
