% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{data}
\name{db}
\alias{db}
\title{Dataset used in Historical Scenarios with Fully Flexible Probabilities.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1083 rows and 9 columns.
}
\usage{
db
}
\description{
Dataset used in Historical Scenarios with Fully Flexible Probabilities.
}
\references{
Meucci, Attilio, Historical Scenarios with Fully Flexible Probabilities
(October 23, 2010). GARP Risk Professional, pp. 47-51, December 2010,
Available at SSRN: https://ssrn.com/abstract=1696802 or http://dx.doi.org/10.2139/ssrn.1696802
}
\seealso{
\code{\link{db_tbl}}
}
\keyword{datasets}
