% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\docType{data}
\name{db_tbl}
\alias{db_tbl}
\title{Dataset used in Historical Scenarios with Fully Flexible Probabilities
(\code{tibble} format).}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1083 rows and 9 columns.
}
\usage{
db_tbl
}
\description{
Dataset used in Historical Scenarios with Fully Flexible Probabilities
(\code{tibble} format).
}
\references{
Meucci, Attilio, Historical Scenarios with Fully Flexible Probabilities
(October 23, 2010). GARP Risk Professional, pp. 47-51, December 2010,
Available at SSRN: \url{https://ssrn.com/abstract=1696802} or \url{http://dx.doi.org/10.2139/ssrn.1696802}
}
\seealso{
\code{\link{db}}
}
\keyword{datasets}
