% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_information.R
\name{double_decay}
\alias{double_decay}
\alias{double_decay.default}
\alias{double_decay.numeric}
\alias{double_decay.matrix}
\alias{double_decay.ts}
\alias{double_decay.xts}
\alias{double_decay.tbl}
\alias{double_decay.data.frame}
\title{Flexible Probabilities using Partial Information}
\usage{
double_decay(x, decay_low, decay_high)

\method{double_decay}{default}(x, decay_low, decay_high)

\method{double_decay}{numeric}(x, decay_low, decay_high)

\method{double_decay}{matrix}(x, decay_low, decay_high)

\method{double_decay}{ts}(x, decay_low, decay_high)

\method{double_decay}{xts}(x, decay_low, decay_high)

\method{double_decay}{tbl}(x, decay_low, decay_high)

\method{double_decay}{data.frame}(x, decay_low, decay_high)
}
\arguments{
\item{x}{A multivariate set of risk-drivers.}

\item{decay_low}{A number with the long half-life (slow decay) for the correlation
matrix.}

\item{decay_high}{A number with the short-life (high decay) for the volatility
structure.}
}
\value{
A vector with the new probabilities distribution.
}
\description{
This function uses entropy-pooling to match different decay factors on the
covariance matrix.
}
\examples{
l_c <- 0.0055
l_s <- 0.0166
ret <- diff(log(EuStockMarkets))
plot(double_decay(ret, l_c, l_s), type = "l")
}
