% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleDecay.R
\name{DoubleDecay}
\alias{DoubleDecay}
\title{Double-Decay Covariance Matrix}
\usage{
DoubleDecay(x, decay_low, decay_high)
}
\arguments{
\item{x}{A set of relevant risk drivers.}

\item{decay_low}{A \code{numeric} value with the low decay (long half-life).}

\item{decay_high}{A \code{numeric} value with the high decay (short half-life).}
}
\value{
A \code{list} with the posterior mean ans sigma.
}
\description{
This function computes the covariance matrix using two different decay factors.
}
\details{
A common practice is to estimate the covariance of the risk drivers using a high
decay (short half-life) for the volatilities and a low decay (long half-life)
for the correlations.
}
\keyword{internal}
