% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_stats.R
\name{empirical_stats}
\alias{empirical_stats}
\alias{empirical_stats.default}
\alias{empirical_stats.numeric}
\alias{empirical_stats.matrix}
\alias{empirical_stats.xts}
\alias{empirical_stats.ts}
\alias{empirical_stats.data.frame}
\alias{empirical_stats.tbl_df}
\title{Summary Statistics for Empirical Distributions}
\usage{
empirical_stats(x, p, level = 0.01)

\method{empirical_stats}{default}(x, p, level = 0.01)

\method{empirical_stats}{numeric}(x, p, level = 0.01)

\method{empirical_stats}{matrix}(x, p, level = 0.01)

\method{empirical_stats}{xts}(x, p, level = 0.01)

\method{empirical_stats}{ts}(x, p, level = 0.01)

\method{empirical_stats}{data.frame}(x, p, level = 0.01)

\method{empirical_stats}{tbl_df}(x, p, level = 0.01)
}
\arguments{
\item{x}{A set of joint-scenarios of risk-drivers.}

\item{p}{A numeric vector with flexible probabilities.}

\item{level}{A number with the desired confidence level.}
}
\value{
A \code{tibble} with 6 rows and the following components:
\itemize{
\item Mu,
\item Std
\item Skew
\item Skew
\item Kurt
\item Var
\item CVaR
}
}
\description{
Computes some of the most commonly used statistics by portfolio managers.
}
\details{
The data in \code{x} and \code{p} are expected to have the same number of rows
(size).
}
\examples{
ret <- diff(log(EuStockMarkets))

# with equal weights
ew  <- rep(1 / nrow(ret), nrow(ret))
empirical_stats(x = ret, p = ew)

# with ffp
exp_smooth <- smoothing(ret, 0.015)
empirical_stats(ret, exp_smooth)
}
