% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_stats.R
\name{empirical_stats}
\alias{empirical_stats}
\alias{empirical_stats.default}
\alias{empirical_stats.double}
\alias{empirical_stats.matrix}
\alias{empirical_stats.xts}
\alias{empirical_stats.ts}
\alias{empirical_stats.data.frame}
\alias{empirical_stats.tbl_df}
\title{Summary Statistics for Empirical Distributions}
\usage{
empirical_stats(x, p, level = 0.01)

\method{empirical_stats}{default}(x, p, level = 0.01)

\method{empirical_stats}{double}(x, p, level = 0.01)

\method{empirical_stats}{matrix}(x, p, level = 0.01)

\method{empirical_stats}{xts}(x, p, level = 0.01)

\method{empirical_stats}{ts}(x, p, level = 0.01)

\method{empirical_stats}{data.frame}(x, p, level = 0.01)

\method{empirical_stats}{tbl_df}(x, p, level = 0.01)
}
\arguments{
\item{x}{A set of joint-scenarios of risk-drivers.}

\item{p}{A numeric vector with flexible probabilities.}

\item{level}{A number with the desired confidence level.}
}
\value{
A table with 6 components:
    \itemize{
      \item{\code{mu}}: A \code{numeric} with the location parameter.
      \item{\code{sd}}: A \code{numeric} with the standard deviation.
      \item{\code{skew}}: A \code{numeric} with the skewness.
      \item{\code{kurtosis}}: A \code{numeric} with the kurtosis.
      \item{\code{VaR}}: A \code{numeric} with the Value-at-Risk.
      \item{\code{CVaR}}: A \code{numeric} with the Expected-Shortfall.
    }
}
\description{
Computes some of the most commonly used statistics by portfolio managers.
}
\details{
The data in \code{x} and \code{p} are expected to have the same number of rows
(size).
}
\examples{
ret <- diff(log(EuStockMarkets))
ew  <- rep(1 / nrow(ret), nrow(ret))

empirical_stats(x = ret, p = ew)
}
