% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens.R
\name{ens}
\alias{ens}
\title{Effective Number of Scenarios}
\usage{
ens(p)
}
\arguments{
\item{p}{An object of the \code{ffp} class.}
}
\value{
A single \code{double}.
}
\description{
Shows how many scenarios are effectively been considered when using flexible
probabilities.
}
\examples{
set.seed(123)
p <- exp_decay(stats::rnorm(100), 0.01)

# ens is smaller than 100
ens(p)
}
