% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_pooling.R
\name{entropy_pooling}
\alias{entropy_pooling}
\title{Numerical Entropy Minimization}
\usage{
entropy_pooling(p, A = NULL, b = NULL, Aeq, beq)
}
\arguments{
\item{p}{A vector of prior probabilities}

\item{A}{The linear inequality constraint (left-hand side)}

\item{b}{The linear inequality constraint (right-hand side)}

\item{Aeq}{The linear equality constraint (left-hand side)}

\item{beq}{The linear equality constraint (right-hand side)}
}
\value{
A vector of posterior probabilities.
}
\description{
This function solves the entropy minimization problem with equality and inequality
constraints. The solution is a vector of posterior probabilities that distorts
the least the prior, given the constraints.
}
\examples{
#
}
\references{
Meucci, Attilio, Fully Flexible Views: Theory and Practice (August 8, 2008).
Fully Flexible Views: Theory and Practice, Risk, Vol. 21, No. 10, pp. 97-102,
October 2008, Available at SSRN: https://ssrn.com/abstract=1213325
}
