% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp-vctrs.R
\name{ffp}
\alias{ffp}
\title{Fully-Flexible Probability Class}
\usage{
ffp(x = double())
}
\arguments{
\item{x}{\itemize{
\item For \code{ffp()}: A numeric vector (must be a probability)
\item For \code{is_ffp()}: An object to test.
\item For \code{as_ffp()}: An object to convert.
}}
}
\value{
A S3 vector of class \code{ffp}.
}
\description{
Functions associated with the \code{ffp} class.
}
\examples{
set.seed(123)
p <- runif(5)
p <- p / sum(p)
ffp(p)
}
