% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_ffp}
\alias{new_ffp}
\alias{validate_ffp}
\alias{ffp}
\alias{is_ffp}
\alias{as_ffp}
\title{Internal ffp class}
\usage{
new_ffp(x, ...)

validate_ffp(x)

ffp(x, ...)

is_ffp(x)

as_ffp(x)
}
\arguments{
\item{x}{The data to be converted to ffp.}

\item{...}{Additional attributes to be passed to ffp.}
}
\description{
Internal ffp class
}
\keyword{internal}
