% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/ffp-vctrs.R
\name{new_ffp_mat}
\alias{new_ffp_mat}
\alias{validate_ffp_mat}
\alias{ffp_mat}
\alias{is_ffp_mat}
\alias{as_ffp_mat}
\title{Internal class ffp_mat}
\usage{
new_ffp_mat(.x)

validate_ffp_mat(.x)

ffp_mat(.x, ...)

is_ffp_mat(.x)

as_ffp_mat(.x)

new_ffp_mat(.x)

validate_ffp_mat(.x)

ffp_mat(.x, ...)

is_ffp_mat(.x)

as_ffp_mat(.x)
}
\description{
Internal class ffp_mat

Internal class ffp_mat
}
\keyword{internal}
