% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_moments.R
\name{fit_to_moments}
\alias{fit_to_moments}
\title{Double-Decay Covariance Matrix by Entropy-Pooling}
\usage{
fit_to_moments(X, m, S)
}
\arguments{
\item{X}{A \code{matrix} of size \code{TxN} with relevant the risk drivers.}

\item{m}{A \code{matrix} of size \code{Nx1} given by \code{Double_Decay}
with risk drivers location.}

\item{S}{A \code{matrix} of size \code{NxN} given by \code{Double_Decay}
with the dispersion matrix of the risk drivers.}
}
\value{
A \code{matrix} vector with the posterior probabilities.
}
\description{
Internal function that computes the flexible probabilities given by
double-decay covariance matrix.
}
\keyword{internal}
