% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_information.R
\name{kernel_normal}
\alias{kernel_normal}
\alias{kernel_normal.default}
\alias{kernel_normal.numeric}
\alias{kernel_normal.matrix}
\alias{kernel_normal.ts}
\alias{kernel_normal.xts}
\alias{kernel_normal.tbl_df}
\alias{kernel_normal.data.frame}
\title{Full Information by Kernel-Damping}
\usage{
kernel_normal(x, mean, sigma)

\method{kernel_normal}{default}(x, mean, sigma)

\method{kernel_normal}{numeric}(x, mean, sigma)

\method{kernel_normal}{matrix}(x, mean, sigma)

\method{kernel_normal}{ts}(x, mean, sigma)

\method{kernel_normal}{xts}(x, mean, sigma)

\method{kernel_normal}{tbl_df}(x, mean, sigma)

\method{kernel_normal}{data.frame}(x, mean, sigma)
}
\arguments{
\item{x}{An univariate set of risk-drivers.}

\item{mean}{A number in which the kernel should be centered.}

\item{sigma}{A number with the uncertainty (volatility) around mean.}
}
\value{
A vector with the new probabilities distribution.
}
\description{
In this framework, historical realizations receive a weight proportional to
the distance from the target mean, which is enveloped
normal kernel with a bandwidth equal sigma.
}
\examples{
ret <- diff(log(EuStockMarkets[ , 1]))
mean <- -0.01 # scenarios around -1\%
sigma <- var(diff(ret))

#plot(kernel_normal(ret, mean, sigma), type = 'l')
}
