% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_density.R
\name{scenario_density}
\alias{scenario_density}
\title{Plot Scenario Densities}
\usage{
scenario_density(x, p)
}
\arguments{
\item{x}{An univariate marginal distribution.}

\item{p}{A vector of fully flexible probabilities.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function is designed make it easy to visualize the impact of a view in the
P&L distribution.
}
\examples{
pnl <- diff(log(EuStockMarkets))[, 1]
p <- smoothing(pnl, 0.005)
scenario_density(pnl, p)

p2 <- double_decay(pnl, 0.01, 0.001)
scenario_density(pnl, p2)
}
