% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_information.R
\name{smoothing}
\alias{smoothing}
\alias{smoothing.default}
\alias{smoothing.numeric}
\alias{smoothing.matrix}
\alias{smoothing.ts}
\alias{smoothing.xts}
\alias{smoothing.data.frame}
\alias{smoothing.tbl}
\title{Full Information by Exponential Smoothing}
\usage{
smoothing(x, lambda)

\method{smoothing}{default}(x, lambda)

\method{smoothing}{numeric}(x, lambda)

\method{smoothing}{matrix}(x, lambda)

\method{smoothing}{ts}(x, lambda)

\method{smoothing}{xts}(x, lambda)

\method{smoothing}{data.frame}(x, lambda)

\method{smoothing}{tbl}(x, lambda)
}
\arguments{
\item{x}{The risk-drivers.}

\item{lambda}{A number with the decay parameter that generated the half-life.}
}
\value{
A \code{tibble} with the new probabilities distribution.
}
\description{
Exponential smoothing twists probabilities by giving relatively more weight
to recent observations at an exponential rate.
}
\details{
The half-life is computed as:

\code{HL = log(2) / lambda}.

For example: log(2) / 0.0166 is approximately 42. So, a parameter \code{lambda} of
0.0166 can be associated with a half-life of two-months.
}
\examples{
library(ggplot2)
# long half_life
long_hl <- smoothing(EuStockMarkets, 0.001)
long_hl

# long half_file
autoplot(long_hl)

# short half_life
short_hl <- smoothing(EuStockMarkets, 0.015)
short_hl
autoplot(short_hl)
}
\seealso{
\code{\link{crisp}} \code{\link{kernel_normal}} \code{\link{half_life}}
}
