% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_copula}
\alias{view_on_copula}
\alias{view_on_copula.default}
\alias{view_on_copula.matrix}
\alias{view_on_copula.xts}
\alias{view_on_copula.tbl_df}
\title{Views on Copulas}
\usage{
view_on_copula(x, simul, p)

\method{view_on_copula}{default}(x, simul, p)

\method{view_on_copula}{matrix}(x, simul, p)

\method{view_on_copula}{xts}(x, simul, p)

\method{view_on_copula}{tbl_df}(x, simul, p)
}
\arguments{
\item{x}{The empirical copula.}

\item{simul}{A simulated target copula.}

\item{p}{An object of the \code{ffp} class.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on copulas for entropy programming.
}
\examples{
# invariants
ret <- diff(log(EuStockMarkets))
cop <- apply(ret, 2, stats::pnorm)
# simulated marginals
n <- nrow(cop)
simul_cop <- cbind(
  runif(n),
  runif(n),
  runif(n),
  runif(n)
)
colnames(simul_cop) <- colnames(EuStockMarkets)

# prior probability distribution
prior <- rep(1 / n, n)

views <- view_on_copula(x = cop, simul = simul_cop, p = prior)
views

#ep <- entropy_pooling(p = prior, Aeq = views$Aeq, beq = views$beq, solver = "nlminb")

#ffp_moments(x = ret, p = ep)
}
