% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_copula}
\alias{view_on_copula}
\alias{view_on_copula.default}
\alias{view_on_copula.matrix}
\alias{view_on_copula.xts}
\alias{view_on_copula.tbl_df}
\title{Views on Copulas}
\usage{
view_on_copula(x, simul, p)

\method{view_on_copula}{default}(x, simul, p)

\method{view_on_copula}{matrix}(x, simul, p)

\method{view_on_copula}{xts}(x, simul, p)

\method{view_on_copula}{tbl_df}(x, simul, p)
}
\arguments{
\item{x}{The empirical copula.}

\item{simul}{A simulated target copula.}

\item{p}{An object of the \code{ffp} class.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on copulas for entropy programming.
}
\examples{
#
}
