% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_correlation}
\alias{view_on_correlation}
\alias{view_on_correlation.default}
\alias{view_on_correlation.numeric}
\alias{view_on_correlation.matrix}
\alias{view_on_correlation.ts}
\alias{view_on_correlation.xts}
\alias{view_on_correlation.tbl_df}
\title{Views on the Correlation Matrix}
\usage{
view_on_correlation(x, cor)

\method{view_on_correlation}{default}(x, cor)

\method{view_on_correlation}{numeric}(x, cor)

\method{view_on_correlation}{matrix}(x, cor)

\method{view_on_correlation}{ts}(x, cor)

\method{view_on_correlation}{xts}(x, cor)

\method{view_on_correlation}{tbl_df}(x, cor)
}
\arguments{
\item{x}{An univariate or a multivariate dataset.}

\item{cor}{A \code{matrix} for the target correlation structure of
the series in \code{x}.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on the correlation matrix for
entropy programming.
}
\examples{
# invariants
ret <- diff(log(EuStockMarkets))

# Assume that a panic event throws correlations to the roof!
co <- matrix(0.95, 4, 4)
diag(co) <- 1
co

prior <- rep(1 / nrow(ret), nrow(ret))

views <- view_on_correlation(x = ret, cor = co)
views

ep <- entropy_pooling(p = prior, Aeq = views$Aeq, beq = views$beq, solver = "nlminb")

# prior correlation
stats::cor(ret)
# posterior correlation
stats::cov2cor(ffp_moments(x = ret, p = ep)$sigma)
}
