% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_covariance}
\alias{view_on_covariance}
\alias{view_on_covariance.default}
\alias{view_on_covariance.numeric}
\alias{view_on_covariance.matrix}
\alias{view_on_covariance.ts}
\alias{view_on_covariance.xts}
\alias{view_on_covariance.tbl_df}
\title{Views on the Covariance Matrix}
\usage{
view_on_covariance(x, mean, sigma)

\method{view_on_covariance}{default}(x, mean, sigma)

\method{view_on_covariance}{numeric}(x, mean, sigma)

\method{view_on_covariance}{matrix}(x, mean, sigma)

\method{view_on_covariance}{ts}(x, mean, sigma)

\method{view_on_covariance}{xts}(x, mean, sigma)

\method{view_on_covariance}{tbl_df}(x, mean, sigma)
}
\arguments{
\item{x}{An univariate or a multivariate dataset.}

\item{mean}{A \code{double} for the location parameter of the series in \code{x}.}

\item{sigma}{A \code{matrix} for the target variance-covariance parameter
of the series in \code{x}.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on variance-covariance matrix for
entropy programming.
}
\examples{
# invariants
ret <- diff(log(EuStockMarkets))
# expected returns
mean <- rep(0, 4)
covs <- stats::cov(ret)

prior <- rep(1 / nrow(ret), nrow(ret))

views <- view_on_covariance(x = ret, mean = mean, sigma = covs)
views

ep <- entropy_pooling(p = prior, Aeq = rbind(1, views$Aeq), beq = c(1, views$beq))

ffp_moments(x = ret, p = ep / sum(ep))
}
\keyword{internal}
