% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_covariance}
\alias{view_on_covariance}
\alias{view_on_covariance.default}
\alias{view_on_covariance.numeric}
\alias{view_on_covariance.matrix}
\alias{view_on_covariance.ts}
\alias{view_on_covariance.xts}
\alias{view_on_covariance.tbl_df}
\title{Views on the Covariance Matrix}
\usage{
view_on_covariance(x, mean, sigma)

\method{view_on_covariance}{default}(x, mean, sigma)

\method{view_on_covariance}{numeric}(x, mean, sigma)

\method{view_on_covariance}{matrix}(x, mean, sigma)

\method{view_on_covariance}{ts}(x, mean, sigma)

\method{view_on_covariance}{xts}(x, mean, sigma)

\method{view_on_covariance}{tbl_df}(x, mean, sigma)
}
\arguments{
\item{x}{An univariate or a multivariate dataset.}

\item{mean}{A \code{double} for the location parameter of the series in \code{x}.}

\item{sigma}{A \code{matrix} for the target variance-covariance parameter
of the series in \code{x}.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on variance-covariance matrix for
entropy programming.
}
\examples{
# invariants (stationarity)
ret <- diff(log(EuStockMarkets))

# Expectations for location and dispersion
mean <- colMeans(ret)
covs <- matrix(0, ncol = 4, nrow = 4) # all assets are uncorrelated

# prior probabilities (usually equal weight scheme)
prior <- rep(1 / nrow(ret), nrow(ret))

# construct the views
views <- view_on_covariance(x = ret, mean = mean, sigma = covs)
views

ep <- entropy_pooling(p = prior, Aeq = views$Aeq, beq = views$beq, solver = "solnl")

# original covariance matrix
stats::cov(ret)
# Posterior covariance matrix
# Parameters are much closer than zero than the original ones
ffp_moments(x = ret, p = ep)$sigma
}
\keyword{internal}
