% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_joint_distribution}
\alias{view_on_joint_distribution}
\alias{view_on_joint_distribution.default}
\alias{view_on_joint_distribution.matrix}
\alias{view_on_joint_distribution.xts}
\alias{view_on_joint_distribution.tbl_df}
\title{Views on the Marginal Distribution}
\usage{
view_on_joint_distribution(x, simul, p)

\method{view_on_joint_distribution}{default}(x, simul, p)

\method{view_on_joint_distribution}{matrix}(x, simul, p)

\method{view_on_joint_distribution}{xts}(x, simul, p)

\method{view_on_joint_distribution}{tbl_df}(x, simul, p)
}
\arguments{
\item{x}{An univariate ou multivariate dataset.}

\item{simul}{An univariate or multivariate dataset.}

\item{p}{An object of the \code{ffp} class.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on the entire marginal distribution for
entropy programming.
}
\details{
\itemize{
\item \code{simul} must have the same number of columns than \code{x}
\item \code{p} should have the same number of rows that \code{simul}.
}
}
\examples{
#
}
