% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_marginal_distribution}
\alias{view_on_marginal_distribution}
\alias{view_on_marginal_distribution.default}
\alias{view_on_marginal_distribution.matrix}
\alias{view_on_marginal_distribution.xts}
\alias{view_on_marginal_distribution.tbl_df}
\title{Views on the Marginal Distribution}
\usage{
view_on_marginal_distribution(x, simul, p, on_mean = TRUE, on_sigma = TRUE)

\method{view_on_marginal_distribution}{default}(x, simul, p, on_mean = TRUE, on_sigma = TRUE)

\method{view_on_marginal_distribution}{matrix}(x, simul, p, on_mean = TRUE, on_sigma = TRUE)

\method{view_on_marginal_distribution}{xts}(x, simul, p, on_mean = TRUE, on_sigma = TRUE)

\method{view_on_marginal_distribution}{tbl_df}(x, simul, p, on_mean = TRUE, on_sigma = TRUE)
}
\arguments{
\item{x}{An univariate ou multivariate dataset.}

\item{simul}{An univariate or multivariate dataset.}

\item{p}{An object of the \code{ffp} class.}

\item{on_mean}{A \code{flag}. Should the constraints be added on the mean?}

\item{on_sigma}{A \code{flag}. Should the constraints be added on the covariance
matrix?}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on the entire marginal distribution.
}
\details{
\itemize{
\item \code{simul} must have the same number of columns than \code{x}
\item \code{p} should have the same number of rows that \code{simul}.
}
}
\examples{
# invariants
ret <- diff(log(EuStockMarkets))
# simulated marginals
n <- nrow(ret)
simul <- cbind(
  stats::rnorm(n) ,
  stats::rnorm(n, 0.01) ,
  stats::rt(n, 2, 0.00005) ,
  stats::rnorm(n, 0.03)
)
colnames(simul) <- colnames(EuStockMarkets)

# prior probability distribution
prior <- rep(1 / n, n)

views <- view_on_marginal_distribution(x = ret, simul = simul / 100, p = prior)
views

ep <- entropy_pooling(p = prior, Aeq = views$Aeq, beq = views$beq, solver = "nlminb")

ffp_moments(x = ret, p = ep)
}
