% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_mean}
\alias{view_on_mean}
\alias{view_on_mean.default}
\alias{view_on_mean.numeric}
\alias{view_on_mean.matrix}
\alias{view_on_mean.ts}
\alias{view_on_mean.xts}
\alias{view_on_mean.tbl_df}
\title{Views on the Location}
\usage{
view_on_mean(x, mean)

\method{view_on_mean}{default}(x, mean)

\method{view_on_mean}{numeric}(x, mean)

\method{view_on_mean}{matrix}(x, mean)

\method{view_on_mean}{ts}(x, mean)

\method{view_on_mean}{xts}(x, mean)

\method{view_on_mean}{tbl_df}(x, mean)
}
\arguments{
\item{x}{An univariate or a multivariate dataset.}

\item{mean}{A \code{double} for the target location parameter of the series in \code{x}.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on the means for entropy programming.
}
\examples{
# invariants
ret <- diff(log(EuStockMarkets))
# expected returns
mean <- rep(0.025, 4)
prior <- rep(1 / nrow(ret), nrow(ret))

views <- view_on_mean(x = ret, mean = mean)
views

ep <- entropy_pooling(p = prior, Aeq = rbind(1, views$Aeq), beq = c(1, views$beq))

ffp_moments(x = ret, p = ep / sum(ep))
}
