% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_rank}
\alias{view_on_rank}
\alias{view_on_rank.default}
\alias{view_on_rank.matrix}
\alias{view_on_rank.ts}
\alias{view_on_rank.xts}
\alias{view_on_rank.tbl_df}
\title{Views on Relative Performance}
\usage{
view_on_rank(x, p, rank)

\method{view_on_rank}{default}(x, p, rank)

\method{view_on_rank}{matrix}(x, p, rank)

\method{view_on_rank}{ts}(x, p, rank)

\method{view_on_rank}{xts}(x, p, rank)

\method{view_on_rank}{tbl_df}(x, p, rank)
}
\arguments{
\item{x}{An univariate ou multivariate dataset.}

\item{p}{An object of the \code{ffp} class.}

\item{rank}{A \code{double} with the asset indexes.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on ranking assets for entropy programming.
}
\examples{
x <- matrix(diff(log(EuStockMarkets)), ncol = 4)
colnames(x) <- colnames(EuStockMarkets)
prior <- as.matrix(rep(1/nrow(x), nrow(x)))
# asset in the first col will outperform the asset in the second col.
views <- view_on_rank(x = x, p = prior, rank = c(2, 1, 4))
views
}
