% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_tail_dependence}
\alias{view_on_tail_dependence}
\alias{view_on_tail_dependence.default}
\alias{view_on_tail_dependence.matrix}
\alias{view_on_tail_dependence.xts}
\alias{view_on_tail_dependence.tbl_df}
\title{Views on Tail Dependence}
\usage{
view_on_tail_dependence(u, tail)

\method{view_on_tail_dependence}{default}(u, tail)

\method{view_on_tail_dependence}{matrix}(u, tail)

\method{view_on_tail_dependence}{xts}(u, tail)

\method{view_on_tail_dependence}{tbl_df}(u, tail)
}
\arguments{
\item{u}{A multivariate copula.}

\item{tail}{A \code{double} with tail index of each asset.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on tail dependence for entropy programming.
}
\examples{
#
}
