% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_tail_dependence}
\alias{view_on_tail_dependence}
\alias{view_on_tail_dependence.default}
\alias{view_on_tail_dependence.matrix}
\alias{view_on_tail_dependence.xts}
\alias{view_on_tail_dependence.tbl_df}
\title{Views on Tail Dependence}
\usage{
view_on_tail_dependence(x, tail)

\method{view_on_tail_dependence}{default}(x, tail)

\method{view_on_tail_dependence}{matrix}(x, tail)

\method{view_on_tail_dependence}{xts}(x, tail)

\method{view_on_tail_dependence}{tbl_df}(x, tail)
}
\arguments{
\item{x}{A multivariate copula.}

\item{tail}{A \code{double} with tail index of each asset.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct views on tail dependence.
}
\examples{
u <- matrix(stats::runif(100 * 2), ncol = 2)
tail <- c(0.25, 0.3)

view_on_tail_dependence(x = u, tail = tail)
}
