% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_volatility}
\alias{view_on_volatility}
\alias{view_on_volatility.default}
\alias{view_on_volatility.numeric}
\alias{view_on_volatility.matrix}
\alias{view_on_volatility.ts}
\alias{view_on_volatility.xts}
\alias{view_on_volatility.tbl_df}
\title{Views on the Volatility}
\usage{
view_on_volatility(x, vol)

\method{view_on_volatility}{default}(x, vol)

\method{view_on_volatility}{numeric}(x, vol)

\method{view_on_volatility}{matrix}(x, vol)

\method{view_on_volatility}{ts}(x, vol)

\method{view_on_volatility}{xts}(x, vol)

\method{view_on_volatility}{tbl_df}(x, vol)
}
\arguments{
\item{x}{An univariate or a multivariate dataset.}

\item{vol}{A \code{double} for the target volatility structure
of the series in \code{x}.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on volatility for entropy programming.
}
\examples{
# invariants
ret <- diff(log(EuStockMarkets))

# expected volalility 30\% higher than average
vol <- apply(ret, 2, stats::sd) * 1.3

# Prior Probabilities
prior <- rep(1 / nrow(ret), nrow(ret))

views <- view_on_volatility(x = ret, vol = vol)
views

ep <- entropy_pooling(p = prior, Aeq = views$Aeq, beq = views$beq, solver = "nlminb")

# desired vol
vol
# posterior vol
pvol <- sqrt(diag(ffp_moments(x = ret, p = ep / sum(ep))$sigma))
pvol
# ratio among prior and posterior vol
pvol / vol
}
