% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_on_volatility}
\alias{view_on_volatility}
\alias{view_on_volatility.default}
\alias{view_on_volatility.numeric}
\alias{view_on_volatility.matrix}
\alias{view_on_volatility.ts}
\alias{view_on_volatility.xts}
\alias{view_on_volatility.tbl_df}
\title{Views on the Volatility}
\usage{
view_on_volatility(x, vol)

\method{view_on_volatility}{default}(x, vol)

\method{view_on_volatility}{numeric}(x, vol)

\method{view_on_volatility}{matrix}(x, vol)

\method{view_on_volatility}{ts}(x, vol)

\method{view_on_volatility}{xts}(x, vol)

\method{view_on_volatility}{tbl_df}(x, vol)
}
\arguments{
\item{x}{An univariate or a multivariate dataset.}

\item{vol}{A \code{double} for the target volatility structure
of the series in \code{x}.}
}
\value{
A \code{list} of the \code{view} class.
}
\description{
Helper to construct constraints on volatility for entropy programming.
}
\examples{
#
}
